Оновлює файли

+++ src/js/03-feedback.js
@@ -2,6 +2,7 @@ import throttle from 'lodash.throttle';
 
 const feedbackFormRef = document.querySelector('.feedback-form');
 const LOCALSTORAGE_KEY = 'feedback-form-state';
+const { email, message } = feedbackFormRef.elements;
 
 checkLocalStorage();
 
@@ -12,8 +13,8 @@ function onFormInput() {
   localStorage.setItem(
     LOCALSTORAGE_KEY,
     JSON.stringify({
-      email: feedbackFormRef.elements.email.value,
-      message: feedbackFormRef.elements.message.value,
+      email: email.value,
+      message: message.value,
     })
   );
 }
@@ -22,16 +23,20 @@ function checkLocalStorage() {
   if (localStorage.getItem(LOCALSTORAGE_KEY)) {
     const savedObj = JSON.parse(localStorage.getItem(LOCALSTORAGE_KEY));
 
-    feedbackFormRef.elements.email.value = savedObj.email;
-    feedbackFormRef.elements.message.value = savedObj.message;
+    email.value = savedObj.email;
+    message.value = savedObj.message;
   }
 }
 
 function onBtnSubmit(e) {
   e.preventDefault();
 
-  console.log(JSON.parse(localStorage.getItem(LOCALSTORAGE_KEY)));
+  if (email.value === '' || message.value === '') {
+    alert('Всі поля повинні бути заповнені!');
+  } else {
+    console.log(JSON.parse(localStorage.getItem(LOCALSTORAGE_KEY)));
 
-  e.currentTarget.reset();
-  localStorage.clear();
+    e.currentTarget.reset();
+    localStorage.clear();
+  }
 }
